// frontend/src/app/api/model/answer-type.ts
/**
 * Checklist
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type AnswerType = 'None' | 'FreeTextAndJustification' | 'FreeText';

export const AnswerType = {

    None: 'None' as AnswerType,

    FreeTextAndJustification: 'FreeTextAndJustification' as AnswerType,

    FreeText: 'FreeText' as AnswerType
};