// frontend/src/app/api/model/question.ts
/**
 * Checklist
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AnswerType } from './answer-type';
import { ConferenceInstance3 } from './conference-instance3';
import { QuestionGroup } from './question-group';


export interface Question { 
    id?: string | null;
    questionGroups: Array<QuestionGroup>;
    conference: Array<ConferenceInstance3>;
    question: string;
    answerType: AnswerType;
}
export namespace Question {
}