// frontend/src/app/api/model/save-checklist-request.ts
/**
 * Checklist
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SaveQuestionRequest } from './save-question-request';


export interface SaveChecklistRequest { 
    uuid?: string | null;
    name: string;
    description: string;
    questionRequests: Array<SaveQuestionRequest>;
}