// frontend/src/app/api/model/save-question-request.ts
/**
 * Checklist
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SavedQuestion2 } from './saved-question2';


export interface SaveQuestionRequest { 
    question: SavedQuestion2;
    originalQuestion?: string | null;
}